requirements.txt
asgiref==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
django-phonenumber-field[phonenumbers]==7.1.0 ; python_version >= "3.9" and python_version < "4.0"
django==4.2.2 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework-simplejwt==5.2.2 ; python_version >= "3.9" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.9" and python_version < "4.0"
drf-spectacular==0.26.2 ; python_version >= "3.9" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
install==1.3.5 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.9" and python_version < "4.0"
phonenumbers==8.13.14 ; python_version >= "3.9" and python_version < "4.0"
pyjwt==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
rest-framework-simplejwt==0.0.2 ; python_version >= "3.9" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.6.3 ; python_version >= "3.9" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.9" and python_version < "4.0"
